requirements.txt
Appium-Python-Client==2.6.1;python_version >= '3.0'
browserstack-local==1.2.2
selene==2.0.0rc8
pytest
python-dotenv
pydantic-settings
allure-pytest